+++ shared/dev-builders/build-gallery-release.js
@@ -125,7 +125,7 @@ var BuildGalleryRelease;
             galleryLinks.push("\n--------------------------"
                 + ("\n### " + ex.folder + "\n")
                 + ("" + ex.description.join('  \n') + '  '));
-            var storyFramesFolder = "" + dev_web_links_js_1.DevWebLinks.repos.galleryReleaseRaw + ex.folder + "/story-frames";
+            var storyFramesFolder = ex.folder + "/story-frames";
             if (!ex.noGifImage) {
                 checkFile("./" + BuildGalleryRelease.SRC_GALLERY_PATH + "/" + ex.folder + "/story-frames/story.gif");
                 galleryLinks.push("\n  "
@@ -136,7 +136,7 @@ var BuildGalleryRelease;
                     + ("\n[![Image](" + storyFramesFolder + "/../" + ex.movieSnapshot + ")]")
                     + ("(" + storyFramesFolder + "/../" + ex.movieViewPage + ")" + '  ' + "\n  "));
             }
-            galleryLinks.push("\nDownload code: [zip](" + dev_web_links_js_1.DevWebLinks.repos.galleryReleaseRaw + ex.folder + "/" + BuildGalleryRelease.EXAMPLE_ZIP_FILE + ")" + '  ' + "\nTry it <a href=\"" + dev_web_links_js_1.DevWebLinks.repos.galleryReleaseRaw + ex.folder + "/index-online.html\">online</a>." + '  ' + "\n" + (ex.usesLive ? '**WARNING** This example requires a live server.  \n' : '  \n') + "\n" + (!ex.teleportable ? '**WARNING** This example doesn\'t supports teleportation.  \n' : '  \n') + "\n    ");
+            galleryLinks.push("\nDownload code: [zip](" + ex.folder + "/" + BuildGalleryRelease.EXAMPLE_ZIP_FILE + ")" + '  ' + "\nTry it <a href=\"" + ex.folder + "/index-online.html\">online</a>." + '  ' + "\n" + (ex.usesLive ? '**WARNING** This example requires a live server.  \n' : '  \n') + "\n" + (!ex.teleportable ? '**WARNING** This example doesn\'t supports teleportation.  \n' : '  \n') + "\n    ");
         });
         var outREADME = fsix_js_1.fsix.readUtf8Sync(BuildGalleryRelease.SRC_GALLERY_PATH + "/README-rel.md")
             + galleryLinks.join('');
@@ -166,16 +166,16 @@ var BuildGalleryRelease;
             // this sytle used in html output of the markdown is designed to be similar
             // to the github markdown rendered in order to have a good simulation of
             // how the user will see the documentation.
-            + (highlightJs ? "\n  <link rel=\"stylesheet\" href=\"node_modules/github.css\">\n  <link rel=\"stylesheet\" href=\"node_modules/font-awesome.css\">\n  <style>\n  body {\n    font-family: -apple-system,BlinkMacSystemFont,\"Segoe UI\",Helvetica,Arial,\n      sans-serif,\"Apple Color Emoji\",\"Segoe UI Emoji\",\"Segoe UI Symbol\";\n      color: #24292e;\n      background-color: white;\n  }\n  pre {\n    background-color: #1b1f230d;\n    padding: 0.2em 0.4em;\n  }\n  #__page {\n     max-width: 980px;\n     padding: 45px;\n     border: 1px solid #ddd;\n     border-bottom-right-radius: 3px;\n     border-bottom-left-radius: 3px;\n     margin-left: 20px;\n  }\n  code {\n    background-color: #1b1f230d;\n    padding: 0.2em 0.4em;\n  }\n  pre code {\n    background-color: transparent;\n    padding: 0;\n  }\n\n  </style>\n  <script src=\"node_modules/highlight.js\"></script>\n  <script>hljs.initHighlightingOnLoad();</script>\n" : '')
+            + (highlightJs ? "\n  <link rel=\"stylesheet\" href=\"node_modules/github.css\">\n  <link rel=\"stylesheet\" href=\"node_modules/font-awesome.css\">\n  <style>\n  body {\n    font-family: -apple-system,BlinkMacSystemFont,\"Segoe UI\",Helvetica,Arial,\n      sans-serif,\"Apple Color Emoji\",\"Segoe UI Emoji\",\"Segoe UI Symbol\";\n      color: #24292e;\n      background-color: white;\n  }\n  pre {\n    background-color: #1b1f230d;\n    padding: 0.2em 0.4em;\n  }\n  #__page {\n     max-width: 980px;\n     padding: 45px;\n     border-bottom-right-radius: 3px;\n     border-bottom-left-radius: 3px;\n     margin-left: 20px;\n  }\n  code {\n    background-color: #1b1f230d;\n    padding: 0.2em 0.4em;\n  }\n  pre code {\n    background-color: transparent;\n    padding: 0;\n  }\n\n  </style>\n  <script src=\"node_modules/highlight.js\"></script>\n  <script>hljs.initHighlightingOnLoad();</script>\n" : '')
             + '</head>\n<body>\n<div id=__page>'
             + markdownCompiler(sourceMarkdown)
             + '</div></body>\n</html>';
         return html;
     }
-    function buildIndexHtml(readMe) {
+    function buildIndexHtml(readMe, gaID) {
         var markdownCompiler = require('marked');
         var highlightJs = require('highlight.js');
-        var indexHtml = markdownToHtml(markdownCompiler, highlightJs, readMe);
+        var indexHtml = integrateHtmlWithWebSite(markdownToHtml(markdownCompiler, highlightJs, readMe), { gaID: gaID }, true);
         sysFs.writeFileSync(BuildGalleryRelease.DST_GALLERY_RELEASE_PATH + "/index.html", indexHtml);
     }
     BuildGalleryRelease.buildIndexHtml = buildIndexHtml;
@@ -219,5 +219,25 @@ var BuildGalleryRelease;
         });
     }
     BuildGalleryRelease.buildGifs = buildGifs;
+    function integrateHtmlWithWebSite(content, opts, isIndex) {
+        var css = "\n<style>\nbody {\n  margin: 0 !important;\n}\n.website-header {\n  background-color: #2C3E50 !important;\n  font-family: \"Lato\", -apple-system, BlinkMacSystemFont, \"Segoe UI\", Roboto, \n  \"Helvetica Neue\", Arial, sans-serif, \"Apple Color Emoji\", \"Segoe UI Emoji\", \"Segoe UI Symbol\";\n  font-size: 16px;\n  font-weight: 400;\n  line-height: 1.5;\n  color: white;\n  text-align: left;\n  padding: 1rem 1rem;\n}\n.website-header a:link, .website-header a:visited {\n  color: white;\n}\n.website-header .separator {\n  padding: 3px;\n}\n  </style>";
+        if (!isIndex) {
+            content = content
+                .replace(/"abeamer\//g, "\"" + opts.onlineLink + "/")
+                .replace(/(<head>)/, "$1\n<title>ABeamer example: [" + opts.folder + "]</title>");
+        }
+        else {
+            content = content
+                .replace(/(<head>)/, "$1\n<title>ABeamer Gallery</title>\n" + css);
+            /*         .replace(/(<body>)/, `$1
+            <div class=website-header><a href="/">Home</a><span class=separator>&gt;</span>Gallery</div>
+            `); */
+        }
+        content = content
+            .replace(/(<head>)/g, '<!-- This file was created to be used online only. -->\n$1')
+            .replace(/<\/head>/, function () { return "\n<script async src=\"https://www.googletagmanager.com/gtag/js?id=" + opts.gaID + "\"></script>\n<script>\n  window.dataLayer = window.dataLayer || [];\n  function gtag(){dataLayer.push(arguments);}\n  gtag('js', new Date());\n  gtag('config', '" + opts.gaID + "');\n</script>\n</head>\n"; });
+        return content;
+    }
+    BuildGalleryRelease.integrateHtmlWithWebSite = integrateHtmlWithWebSite;
 })(BuildGalleryRelease = exports.BuildGalleryRelease || (exports.BuildGalleryRelease = {}));
 //# sourceMappingURL=build-gallery-release.js.map